fix: NPE in initiateHoverRequest (this.cancellationSupport is null)
Fixes #1104
